Created scroodle main file
Created init_connection function, which initialized the first connection with the Moodle website
and log ins to user profile.
The function retrieves the codes for the current semester courses.
Yet to figure out how to calculate "current" semester.
Username and password are to be saved in a different file, my_creds.py.