Program now receives username and password as input
Scroodle now begins by asking for the user's username and password. It then suggets the user as a first-timer to save the username, for future use.
If no, the program proceeds normally. Else, the file my_creds.ini is created, and saves the username under section CREDENTIALS.
Password is not suggested to be saved.
.gitignore was also modified: my_creds.py => my_creds.ini.