Fixed main bug -- program runs
The main bug of scroodle, since the new semester (first semester of 2019) has started, was apparently that in one of my courses, I've had no updates in the forum. Thus, RoboBrowser could'nt have found a table in the forum of this course, and raised an error.
The program now skips these 'empty' course forums.